Move date format to sub-page
