+++ assets/stylesheets/_components.scss
@@ -318,6 +318,7 @@
 @import 'my-sites/site-settings/action-panel/style';
 @import 'my-sites/site-settings/custom-content-types/style';
 @import 'my-sites/site-settings/comment-display-settings/style';
+@import 'my-sites/site-settings/date-time-format/style';
 @import 'my-sites/site-settings/delete-site/style';
 @import 'my-sites/site-settings/delete-site-options/style';
 @import 'my-sites/site-settings/delete-site-warning-dialog/style';
