+++ client/my-sites/site-settings/date-time-format/date-time-format-options.jsx
@@ -3,6 +3,8 @@
  */
 import React, { Component } from 'react';
 import { localize } from 'i18n-calypso';
+import { includes, startsWith } from 'lodash';
+import classNames from 'classnames';
 
 /**
  * Internal dependencies
@@ -10,12 +12,82 @@ import { localize } from 'i18n-calypso';
 import Button from 'components/button';
 import Card from 'components/card';
 import FormFieldset from 'components/forms/form-fieldset';
+import FormInput from 'components/forms/form-text-input';
 import FormLabel from 'components/forms/form-label';
+import FormRadio from 'components/forms/form-radio';
 import FormSelect from 'components/forms/form-select';
 import SectionHeader from 'components/section-header';
+import { phpToMomentDatetimeFormat } from 'lib/formatting';
 import wrapSettingsForm from '../wrap-settings-form';
 
 export class DateTimeFormatOptions extends Component {
+	dateFormatOption() {
+		const {
+			fields: { date_format, timezone_string },
+			handleRadio,
+			isRequestingSettings,
+			moment,
+			onChangeField,
+			translate,
+		} = this.props;
+
+		const defaultFormats = [ 'F j, Y', 'Y-m-d', 'm/d/Y', 'd/m/Y' ];
+		const isCustomFormat = ! includes( defaultFormats, date_format );
+		const today = startsWith( timezone_string, 'UTC' )
+			? moment().utcOffset( timezone_string.substring( 3 ) * 60 )
+			: moment.tz( timezone_string );
+
+		const customFieldClasses = classNames(
+			'date-time-format__custom-field',
+			{ 'is-custom': isCustomFormat }
+		);
+
+		return (
+			<FormFieldset>
+				<FormLabel>
+					{ translate( 'Date Format' ) }
+				</FormLabel>
+				{ defaultFormats.map( ( format, key ) =>
+					<FormLabel key={ key }>
+						<FormRadio
+							checked={ format === date_format }
+							disabled={ isRequestingSettings }
+							name="date_format"
+							onChange={ handleRadio }
+							value={ format }
+						/>
+						<span>{ today.format( phpToMomentDatetimeFormat( format ) ) }</span>
+					</FormLabel>
+				) }
+				<FormLabel className={ customFieldClasses }>
+					<FormRadio
+						checked={ isCustomFormat }
+						disabled={ isRequestingSettings }
+						name="date_format"
+						onChange={ handleRadio }
+						value={ date_format }
+					/>
+					<span>
+						{ translate( 'Custom' ) }
+						<FormInput
+							disabled={ isRequestingSettings }
+							name="date_format_custom"
+							onChange={ onChangeField( 'date_format' ) }
+							type="text"
+							value={ date_format || '' }
+						/>
+						<span className="date-time-format__custom-preview">
+							{ isCustomFormat && date_format
+								? today.format( phpToMomentDatetimeFormat( date_format ) )
+								: ''
+							}
+						</span>
+				</span>
+				</FormLabel>
+			</FormFieldset>
+		);
+	}
+
 	startOfWeekOption() {
 		const {
 			fields: { start_of_week },
@@ -81,6 +153,7 @@ export class DateTimeFormatOptions extends Component {
 				</SectionHeader>
 				<Card>
 					<form>
+						{ this.dateFormatOption() }
 						{ this.startOfWeekOption() }
 					</form>
 				</Card>
@@ -92,8 +165,9 @@ export class DateTimeFormatOptions extends Component {
 export default wrapSettingsForm( settings => {
 	const defaultSettings = {
 		date_format: '',
-		time_format: '',
 		start_of_week: 0,
+		time_format: '',
+		timezone_string: '',
 	};
 
 	if ( ! settings ) {
@@ -102,9 +176,23 @@ export default wrapSettingsForm( settings => {
 
 	const formSettings = {
 		date_format: settings.date_format,
-		time_format: settings.time_format,
 		start_of_week: settings.start_of_week,
+		time_format: settings.time_format,
+		timezone_string: settings.timezone_string,
 	};
+
+	// handling `gmt_offset` and `timezone_string` values
+	const gmt_offset = settings.gmt_offset;
+
+	if (
+		! settings.timezone_string &&
+		typeof gmt_offset === 'string' &&
+		gmt_offset.length
+	) {
+		formSettings.timezone_string = 'UTC' +
+			( /\-/.test( gmt_offset ) ? '' : '+' ) +
+			gmt_offset;
+	}
 
 	return formSettings;
 } )( localize( DateTimeFormatOptions ) );
