+++ client/my-sites/site-settings/date-time-format/style.scss
@@ -1,0 +1,18 @@
+.date-time-format__custom-field {
+	.form-radio {
+		display: none;
+		margin-top: 12px;
+	}
+	&.is-custom .form-radio {
+		display: block;
+	}
+
+	.form-text-input {
+		margin: 0 12px;
+		width: 100px;
+	}
+
+	.date-time-format__custom-preview {
+		white-space: nowrap;
+	}
+}
