+++ client/my-sites/site-settings/form-general.jsx
@@ -423,77 +423,6 @@ class SiteSettingsFormGeneral extends Component {
 		);
 	}
 
-	dateFormatOption() {
-		if ( ! config.isEnabled( 'manage/site-settings/date-time-format' ) ) {
-			return null;
-		}
-
-		const {
-			fields: { date_format, timezone_string },
-			handleRadio,
-			isRequestingSettings,
-			moment,
-			onChangeField,
-			translate,
-		} = this.props;
-
-		const defaultFormats = [ 'F j, Y', 'Y-m-d', 'm/d/Y', 'd/m/Y' ];
-		const isCustomFormat = ! includes( defaultFormats, date_format );
-		const today = startsWith( timezone_string, 'UTC' )
-			? moment().utcOffset( timezone_string.substring( 3 ) * 60 )
-			: moment.tz( timezone_string );
-
-		const customFieldClasses = classNames(
-			'site-settings__date-time-format-custom',
-			{ 'is-custom': isCustomFormat }
-		);
-
-		return (
-			<FormFieldset>
-				<FormLabel>
-					{ translate( 'Date Format' ) }
-				</FormLabel>
-				{ defaultFormats.map( ( format, key ) =>
-					<FormLabel key={ key }>
-						<FormRadio
-							checked={ format === date_format }
-							disabled={ isRequestingSettings }
-							name="date_format"
-							onChange={ handleRadio }
-							value={ format }
-						/>
-						<span>{ today.format( phpToMomentDatetimeFormat( format ) ) }</span>
-					</FormLabel>
-				) }
-				<FormLabel className={ customFieldClasses }>
-					<FormRadio
-						checked={ isCustomFormat }
-						disabled={ isRequestingSettings }
-						name="date_format"
-						onChange={ handleRadio }
-						value={ date_format }
-					/>
-					<span>
-						{ translate( 'Custom' ) }
-						<FormInput
-							disabled={ isRequestingSettings }
-							name="date_format_custom"
-							onChange={ onChangeField( 'date_format' ) }
-							type="text"
-							value={ date_format || '' }
-						/>
-						<span className="site-settings__date-time-format-custom-preview">
-							{ isCustomFormat && date_format
-								? today.format( phpToMomentDatetimeFormat( date_format ) )
-								: ''
-							}
-						</span>
-				</span>
-				</FormLabel>
-			</FormFieldset>
-		);
-	}
-
 	timeFormatOption() {
 		if ( ! config.isEnabled( 'manage/site-settings/date-time-format' ) ) {
 			return null;
